requirements.txt
numpy
pandas
scikit-learn
pydantic
uvicorn
fastapi
python-dotenv
openai
pinecone-client
sqlalchemy
cloud-sql-python-connector[pg8000]
duckdb
duckdb-engine
